finalized QueryNetworkStateService and added documentation around the mirrored grpc classes.
