+++ src/main/kotlin/com/zepben/evolve/streaming/mutations/UpdateNetworkStateService.kt
@@ -12,6 +12,7 @@ import com.zepben.evolve.cim.iec61970.base.core.PhaseCode
 import com.zepben.evolve.conn.grpc.GrpcException
 import com.zepben.evolve.services.common.translator.toLocalDateTime
 import com.zepben.evolve.services.common.translator.toTimestamp
+import com.zepben.evolve.streaming.data.*
 import com.zepben.protobuf.ns.SetCurrentStatesRequest
 import com.zepben.protobuf.ns.SetCurrentStatesResponse
 import com.zepben.protobuf.ns.SetCurrentStatesResponse.StatusCase
